requirements.txt
flask
python-docx
pdfminer.six
scikit-learn
